Increase timeout for some canvas tests.
This allows these tests to pass on debug builds. (On my computer, they
take around 30s or so to render, and the default timeout is 10s.)

I'm really not comfortable with the test manifest format, so I probably
really screwed this up somehow... but it appears to work.

This should probably also be extended to
2dcontext/drawing-images-to-the-canvas/*.